refactor: replace litellm with direct SDK providers
- Create _provider_base.py: shared retry/validation/logging base (DirectProvider)
- anthropic.py: direct anthropic SDK with cache_control on system prompts
- openai.py: direct openai SDK (chat.completions.create)
- openrouter.py: openai SDK at OpenRouter base URL with HTTP-Referer/X-Title headers
- google.py: google-genai SDK with function-calling and retry support
- registry.py: remove litellm_prefix from ProviderSpec; qualify_model_name strips prefix
- tokens.py: remove litellm.token_counter; offline estimator for non-Anthropic
- rate_limited.py: replace litellm exception types with direct SDK exceptions
- Delete litellm_provider.py
- tests: update mocks to chat.completions.create, remove litellm-specific tests
