+++ src/osoji/llm/openrouter.py
@@ -1,10 +1,25 @@
-"""OpenRouter provider compatibility wrapper backed by LiteLLM."""
+"""OpenRouter provider — openai SDK pointed at OpenRouter's API endpoint."""
 
-from .litellm_provider import LiteLLMProvider
+from __future__ import annotations
+
+import os
+
+import openai
+
+from .openai import OpenAIProvider
 
 
-class OpenRouterProvider(LiteLLMProvider):
-    """OpenRouter provider using the shared LiteLLM adapter."""
+class OpenRouterProvider(OpenAIProvider):
+    """OpenRouter provider using the openai SDK with OpenRouter base URL."""
 
-    def __init__(self) -> None:
-        super().__init__("openrouter")
+    _PROVIDER_NAME = "openrouter"
+
+    def _make_client(self, api_key: str) -> openai.AsyncOpenAI:
+        return openai.AsyncOpenAI(
+            api_key=api_key,
+            base_url="https://openrouter.ai/api/v1",
+            default_headers={
+                "HTTP-Referer": os.environ.get("OSOJI_HTTP_REFERER", "https://osoji.ai"),
+                "X-Title": os.environ.get("OSOJI_APP_TITLE", "osoji"),
+            },
+        )
