+++ src/osoji/llm/registry.py
@@ -6,6 +6,10 @@ from dataclasses import dataclass
 
 DEFAULT_PROVIDER = "anthropic"
 
+# Known provider prefixes a caller might include in a model string (e.g. "openai/gpt-4").
+# These are stripped before passing the model name to a direct SDK.
+_KNOWN_PREFIXES = {"anthropic", "openai", "gemini", "google", "openrouter", "claude-code"}
+
 
 @dataclass(frozen=True)
 class ProviderSpec:
@@ -13,7 +17,6 @@ class ProviderSpec:
 
     name: str
     display_name: str
-    litellm_prefix: str
     api_key_env: str
     rate_limit_name: str
     requires_explicit_model: bool
@@ -23,7 +26,6 @@ _PROVIDER_SPECS: dict[str, ProviderSpec] = {
     "anthropic": ProviderSpec(
         name="anthropic",
         display_name="Anthropic",
-        litellm_prefix="anthropic",
         api_key_env="ANTHROPIC_API_KEY",
         rate_limit_name="anthropic",
         requires_explicit_model=False,
@@ -31,7 +33,6 @@ _PROVIDER_SPECS: dict[str, ProviderSpec] = {
     "openai": ProviderSpec(
         name="openai",
         display_name="OpenAI",
-        litellm_prefix="openai",
         api_key_env="OPENAI_API_KEY",
         rate_limit_name="openai",
         requires_explicit_model=True,
@@ -39,7 +40,6 @@ _PROVIDER_SPECS: dict[str, ProviderSpec] = {
     "google": ProviderSpec(
         name="google",
         display_name="Google Gemini",
-        litellm_prefix="gemini",
         api_key_env="GEMINI_API_KEY",
         rate_limit_name="google",
         requires_explicit_model=False,
@@ -47,7 +47,6 @@ _PROVIDER_SPECS: dict[str, ProviderSpec] = {
     "openrouter": ProviderSpec(
         name="openrouter",
         display_name="OpenRouter",
-        litellm_prefix="openrouter",
         api_key_env="OPENROUTER_API_KEY",
         rate_limit_name="openrouter",
         requires_explicit_model=False,
@@ -55,27 +54,20 @@ _PROVIDER_SPECS: dict[str, ProviderSpec] = {
     "claude-code": ProviderSpec(
         name="claude-code",
         display_name="Claude Code CLI",
-        litellm_prefix="",
         api_key_env="",
         rate_limit_name="claude-code",
         requires_explicit_model=False,
     ),
 }
 
-_KNOWN_MODEL_PREFIXES = {
-    spec.litellm_prefix for spec in _PROVIDER_SPECS.values() if spec.litellm_prefix
-} | set(_PROVIDER_SPECS)
-
 
 def provider_names() -> tuple[str, ...]:
     """Return the supported provider names."""
-
     return tuple(sorted(_PROVIDER_SPECS))
 
 
 def normalize_provider_name(name: str | None) -> str:
     """Normalize and validate a provider name."""
-
     normalized = (name or DEFAULT_PROVIDER).strip().lower()
     if normalized not in _PROVIDER_SPECS:
         valid = ", ".join(provider_names())
@@ -85,29 +77,23 @@ def normalize_provider_name(name: str | None) -> str:
 
 def get_provider_spec(name: str | None) -> ProviderSpec:
     """Return metadata for a provider."""
-
     return _PROVIDER_SPECS[normalize_provider_name(name)]
 
 
 def qualify_model_name(provider: str, model: str) -> str:
-    """Add the LiteLLM provider prefix unless the model is already qualified."""
+    """Strip any known provider prefix from a model name.
 
-    spec = get_provider_spec(provider)
-    stripped = model.strip()
-    if "/" in stripped:
-        prefix = stripped.split("/", 1)[0].lower()
-        if prefix in _KNOWN_MODEL_PREFIXES:
-            return stripped
-    if not spec.litellm_prefix:
-        return stripped
-    return f"{spec.litellm_prefix}/{stripped}"
+    Direct SDKs don't use litellm-style prefixes like 'openai/gpt-4'.
+    This normalizes model strings that callers may have already qualified.
+    """
+    return strip_provider_prefix(provider, model)
 
 
 def strip_provider_prefix(provider: str, model: str) -> str:
-    """Strip a provider prefix from a model string when present."""
-
-    spec = get_provider_spec(provider)
-    for prefix in (spec.litellm_prefix, spec.name):
-        if model.startswith(f"{prefix}/"):
-            return model[len(prefix) + 1 :]
-    return model
+    """Strip a known provider prefix from a model string when present."""
+    stripped = model.strip()
+    if "/" in stripped:
+        prefix = stripped.split("/", 1)[0].lower()
+        if prefix in _KNOWN_PREFIXES:
+            return stripped.split("/", 1)[1]
+    return stripped
